+++ src/main/kotlin/kotools/types/number/PositiveInt.kt
@@ -90,40 +90,6 @@ public constructor(override val value: Int) : KotoolsInt {
 
     // ---------- Binary operations ----------
 
-    /** Adds the [other] value to this [value]. */
-    public infix operator fun plus(other: PositiveInt): PositiveInt =
-        PositiveInt(plus(other.value))
-
-    /** Adds the [other] value to this [value]. */
-    public infix operator fun plus(
-        other: StrictlyPositiveInt
-    ): StrictlyPositiveInt = StrictlyPositiveInt(plus(other.value))
-
-    /** Subtracts the [other] value from this [value]. */
-    public infix operator fun minus(other: NegativeInt): PositiveInt =
-        PositiveInt(minus(other.value))
-
-    /** Subtracts the [other] value from this [value]. */
-    public infix operator fun minus(
-        other: StrictlyNegativeInt
-    ): StrictlyPositiveInt = StrictlyPositiveInt(minus(other.value))
-
-    /** Multiplies this [value] by the [other] value. */
-    public infix operator fun times(other: PositiveInt): PositiveInt =
-        PositiveInt(times(other.value))
-
-    /** Multiplies this [value] by the [other] value. */
-    public infix operator fun times(other: StrictlyPositiveInt): PositiveInt =
-        PositiveInt(times(other.value))
-
-    /** Multiplies this [value] by the [other] value. */
-    public infix operator fun times(other: NegativeInt): NegativeInt =
-        NegativeInt(times(other.value))
-
-    /** Multiplies this [value] by the [other] value. */
-    public infix operator fun times(other: StrictlyNegativeInt): NegativeInt =
-        NegativeInt(times(other.value))
-
     /**
      * Divides this [value] by [other], truncating the result to an integer that
      * is closer to `0`.
@@ -131,14 +97,14 @@ public constructor(override val value: Int) : KotoolsInt {
      */
     @Throws(ArithmeticException::class)
     public infix operator fun div(other: PositiveInt): PositiveInt =
-        PositiveInt(div(other.value))
+        div(other.value).toPositiveInt()
 
     /**
      * Divides this [value] by [other], truncating the result to an integer that
      * is closer to `0`.
      */
     public infix operator fun div(other: StrictlyPositiveInt): PositiveInt =
-        PositiveInt(div(other.value))
+        div(other.value).toPositiveInt()
 
     /**
      * Divides this [value] by [other], truncating the result to an integer that
@@ -147,14 +113,14 @@ public constructor(override val value: Int) : KotoolsInt {
      */
     @Throws(ArithmeticException::class)
     public infix operator fun div(other: NegativeInt): NegativeInt =
-        NegativeInt(div(other.value))
+        div(other.value).toNegativeInt()
 
     /**
      * Divides this [value] by [other], truncating the result to an integer that
      * is closer to `0`.
      */
     public infix operator fun div(other: StrictlyNegativeInt): NegativeInt =
-        NegativeInt(div(other.value))
+        div(other.value).toNegativeInt()
 
     // ---------- Conversions ----------